[ci-app] reorg util function based on feedback
