+++ packages/datadog-plugin-jest/test/index.spec.js
@@ -13,6 +13,7 @@ const {
   ERROR_STACK,
   TEST_PARAMETERS
 } = require('../../dd-trace/src/plugins/util/test')
+const { getFormattedJestTestParameters } = require('../src/util')
 
 describe('Plugin', () => {
   let tracer
@@ -21,6 +22,25 @@ describe('Plugin', () => {
   const TEST_NAME = 'test_name'
   const TEST_SUITE = 'test-file.js'
   const BUILD_SOURCE_ROOT = '/source-root'
+
+  describe('getFormattedJestTestParameters', () => {
+    it('returns formatted parameters for arrays', () => {
+      const result = getFormattedJestTestParameters([[[1, 2], [3, 4]]])
+      expect(result).to.eql([[1, 2], [3, 4]])
+    })
+    it('returns formatted parameters for strings', () => {
+      const result = getFormattedJestTestParameters([['\n    a    | b    | expected\n    '], 1, 2, 3, 3, 5, 8, 0, 1, 1])
+      expect(result).to.eql([{ a: 1, b: 2, expected: 3 }, { a: 3, b: 5, expected: 8 }, { a: 0, b: 1, expected: 1 }])
+    })
+    it('does not crash for invalid inputs', () => {
+      const resultUndefined = getFormattedJestTestParameters(undefined)
+      const resultEmptyArray = getFormattedJestTestParameters([])
+      const resultObject = getFormattedJestTestParameters({})
+      expect(resultEmptyArray).to.eql(undefined)
+      expect(resultUndefined).to.eql(undefined)
+      expect(resultObject).to.eql(undefined)
+    })
+  })
 
   withVersions(plugin, ['jest-environment-node', 'jest-environment-jsdom'], (version, moduleName) => {
     afterEach(() => {
